Split up `test.rs` in generic-event pallet
